2.RLAC-IAChronicleSegmenter/pyproject.toml
[project]
name = "2-backend-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "numpy>=2.4.4",
    "requests>=2.33.1",
    "scipy>=1.17.1",
    "openai-whisper>=20231117",
    "flask-socketio>=5.6.1",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.9",
    "gevent>=24.2.1",
    "gevent-websocket>=0.10.1",
    "schedule>=1.2.0",
]
